# feat(search_family): Speed up merging of index results