--- a/src/core/search/base.h
+++ b/src/core/search/base.h
@@ -126,6 +126,8 @@ inline size_t GetHighestPowerOfTwo(size_t n) {
 }  // namespace details
 
 template <typename Iterator> void BasicSeekGE(DocId min_doc_id, const Iterator& end, Iterator* it) {
+  using Category = typename std::iterator_traits<Iterator>::iterator_category;
+
   auto less_than_min_doc_id = [&](const auto& value) {
     using T = std::decay_t<decltype(value)>;
     if constexpr (std::is_same_v<T, DocId>) {
@@ -135,13 +137,15 @@ template <typename Iterator> void BasicSeekGE(DocId min_doc_id, const Iterator& 
     }
   };
 
-  size_t length = std::distance(*it, end);
-  for (size_t step = details::GetHighestPowerOfTwo(length); step > 0; step >>= 1) {
-    if (step < length) {
-      auto next_it = *it + step;
-      if (less_than_min_doc_id(*next_it)) {
-        *it = next_it;
-        length -= step;
+  if constexpr (std::is_base_of_v<std::random_access_iterator_tag, Category>) {
+    size_t length = std::distance(*it, end);
+    for (size_t step = details::GetHighestPowerOfTwo(length); step > 0; step >>= 1) {
+      if (step < length) {
+        auto next_it = *it + step;
+        if (less_than_min_doc_id(*next_it)) {
+          *it = next_it;
+          length -= step;
+        }
       }
     }
   }