--- a/src/core/search/compressed_sorted_set.cc
+++ b/src/core/search/compressed_sorted_set.cc
@@ -206,7 +206,12 @@ std::pair<CompressedSortedSet, CompressedSortedSet> CompressedSortedSet::Split()
 
   // Move iterator to middle position and save size of diffs tail
   auto it = begin();
-  std::advance(it, size_ / 2);
+  std::advance(it, (size_ - 1) / 2);
+
+  // Save last value in the first set
+  tail_value_ = *it;
+  ++it;
+
   size_t keep_bytes = it.last_read_.data() - diffs_.data();
 
   // Copy second half into second set
@@ -215,7 +220,6 @@ std::pair<CompressedSortedSet, CompressedSortedSet> CompressedSortedSet::Split()
 
   // Erase diffs tail
   diffs_.resize(keep_bytes);
-  tail_value_ = std::nullopt;
   size_ -= second.Size();
 
   return std::make_pair(std::move(*this), std::move(second));