Add block-level jumps to the BlockList, as well as power-of-two-length jumps based for base vector iterators.
This optimization speeds up merging for all index types, not just numeric indexes.

Before:
| Benchmark                                                  | Time (ns)  | CPU (ns)   | Iterations |
|------------------------------------------------------------|------------|------------|------------|
| BM_SearchNumericAndTagIndexes/num_docs:10000               | 12752      | 12751      | 5392       |
| BM_SearchNumericAndTagIndexes/num_docs:100000              | 1407975    | 1407833    | 497        |
| BM_SearchNumericAndTagIndexes/num_docs:1000000             | 11055515   | 11054017   | 64         |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:10000        | 154778     | 154766     | 4535       |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:100000       | 2084997    | 2084989    | 335        |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:1000000      | 40143853   | 40141445   | 17         |


After:
| Benchmark                                                  | Time (ns)  | CPU (ns)   | Iterations |
|------------------------------------------------------------|------------|------------|------------|
| BM_SearchNumericAndTagIndexes/num_docs:10000               | 19200      | 19197      | 35978      |
| BM_SearchNumericAndTagIndexes/num_docs:100000              | 35147      | 35141      | 19515      |
| BM_SearchNumericAndTagIndexes/num_docs:1000000             | 10574950   | 10571685   | 65         |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:10000        | 32930      | 32926      | 21365      |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:100000       | 37653      | 37652      | 18845      |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:1000000      | 31691981   | 31686310   | 23         |