@BagritsevichStepan 2025-08-01
Some updates regarding performance:
With try seek I succeed to speed up it in some cases:

Before this PR:
```

```

Always try long steps:
| Benchmark                                             | Time        | CPU         | Iterations |
|--------------------------------------------------------|-------------|-------------|------------|
| BM_SearchMergeEqualSets/num_docs:100                  | 5692 ns     | 5691 ns     | 123016     |
| BM_SearchMergeEqualSets/num_docs:1000                 | 15690 ns    | 15685 ns    | 44832      |
| BM_SearchMergeEqualSets/num_docs:10000                | 105776 ns   | 105751 ns   | 6565       |
| BM_SearchMergeEqualSets/num_docs:100000               | 1007790 ns  | 1007719 ns  | 693        |
| BM_SearchMergeEqualSets/num_docs:1000000              | 34296912 ns | 34293059 ns | 21         |

After try_seek:
| Benchmark                                             | Time        | CPU         | Iterations |
|--------------------------------------------------------|-------------|-------------|------------|
| BM_SearchMergeEqualSets/num_docs:100                  | 5078 ns     | 5077 ns     | 137744     |
| BM_SearchMergeEqualSets/num_docs:1000                 | 8960 ns     | 8959 ns     | 78013      |
| BM_SearchMergeEqualSets/num_docs:10000                | 46215 ns    | 46208 ns    | 15131      |
| BM_SearchMergeEqualSets/num_docs:100000               | 421455 ns   | 421380 ns   | 1664       |
| BM_SearchMergeEqualSets/num_docs:1000000              | 22504371 ns | 22501089 ns | 31         |